+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next release
+
+- Send close connection once the JWT token expires (if channel is open with a token using the `exp` claim).
+
 ## 0.6.1.0
 
 - Add capability to unset `PGWS_ROOT_PATH` to disable static file serving.